requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
python-dotenv==1.0.0
PyJWT==2.8.0
Werkzeug==3.0.1
Flask-Limiter==3.5.0
pydantic<2
Flask-Pydantic
boto3==1.34.94
sendgrid==6.11.0